Adaugare functionalitati in PageInterpreter
Issue #14
Am adaugat functionalitati in `isPageCorrect` si `formatPages`
